sleep: new option: user provided hibernation time [WIP]
Adds a new option where the user can provide a hibernation time range, e.g.,
22:00..06:00.

If the system is within the time range, then it will suspend-then-hibernate as
before. If it is not whitn that time range, then the hibernation is scheduled
for the start of the hibernation time range.